src/cloud_function/cf_customers/requirements.txt
Faker==37.1.0
polars==1.29.0
google-cloud-secret-manager==2.23.3